Fix bounding box computation.
To better match the outputs of the skeleton code, we now expand the
shortest edge of the bouding box in the positive direction (and thus not
symmetrically). This leads to a jump in the MSE from 190 to 139. Some
simple statistics on the dataset indeed show that the keypoint positions
are indeed biased towards zero.